Add extra title field
The dateFormat and timeFormat fields don't seem to be doing what I
thought they'd do, so maybe adding an explicit title field that
functions exactly the same way will work (doubtful, though).

All this just to get a readable string in the CMS for the journal
entries because that has to be difficult, for some reason.