Aaaaand we're back to just a string for the title
I don't understand the documentation well enough to figure out how to
make this work without this extra step. Ideally, the posts would be
identified in the CMS UI by the value from the datetime field, but I
keep getting React errors for some reason.

Oh the joys of working with fledgling technology.