Add markdown filter
We need our own markdown filter so the markdown in our recipe
frontmatter is correctly parsed as markdown so the nunjucks markdown
parser can compile the markdown.

I think. Something like that, anyway.

Markdown in the recipe frontmatter was being rendered as-is, so we need
to add another level of markdowning.

It's turtles all the way down.