Migrate to single 'tags' page
It doesn't look like the "simple tag pages" from 11ty can handle the
sort of complex set up I was attempting (whether that's a problem with
me or the software is a matter of debate).

This change moves us to a standard `/tags/{{ tag }}` path, and we'll use
other, hand-rolled pages (like the existing `/recipes/` page) to perform
the advanced arranging and presenting.